+++ magpie/api/schemas.py
@@ -468,43 +468,120 @@ class QueryRequestSchemaAPI(colander.MappingSchema):
                                     "This formatting alternative by query parameter overrides the Accept header.")
 
 
-QueryEffectivePermissions = colander.SchemaNode(
-    colander.Boolean(), name="effective", default=False, missing=colander.drop,
-    description="Obtain user's effective permissions resolved with corresponding service inheritance functionality. "
-                "(Note: Group inheritance is enforced regardless of 'inherited' query parameter values.)")
-QueryInheritGroupsPermissions = colander.SchemaNode(
-    colander.Boolean(), name="inherited", default=False, missing=colander.drop,
-    description="Include the user's groups memberships inheritance to retrieve all possible permissions. "
-                "(Note: Duplicate, redundant and even conflicting permissions can be obtained when considering "
-                "multiple applied permissions individually applied from different group memberships or for the user "
-                "itself. See 'resolve' query parameter to reduce the set into a single highest priority permission).")
-QueryResolvedUserGroupsPermissions = colander.SchemaNode(
-    colander.Boolean(), name="resolve", default=False, missing=colander.drop,
-    description="Combines corresponding direct user and groups inherited permissions into one, and locally resolves "
-                "for every resource the applicable permission modifiers considering group precedence and priorities. "
-                "(Note: Group permissions retrieval is enforced when using this option regardless of 'inherited' query "
-                "parameter since they are required to perform resolution).")
-QueryFilterResources = colander.SchemaNode(
-    colander.Boolean(), name="filtered", default=False, missing=colander.drop,
-    description="Filter returned resources only where user has permissions on, either directly or inherited by groups "
-                "according to other query parameters. Otherwise (default), return all existing resources "
-                "with empty permission sets when user has no permission on them. Filtered view is enforced for "
-                "non-admin request user.")
-QueryCascadeResourcesPermissions = colander.SchemaNode(
-    colander.Boolean(), name="cascade", default=False, missing=colander.drop,
-    description="Display all services that has at least one permission at any level in his hierarchy "
-                "(including all children resources). Otherwise (default), only returns services that have permissions "
-                "explicitly set on them, ignoring permissions set on children resources.")
-QueryFlattenServices = colander.SchemaNode(
-    colander.Boolean(), name="flatten", default=False, missing=colander.drop,
-    description="Return elements as a flattened list of JSON objects instead of default response format. "
-                "Default is a nested JSON of service-type keys with children service-name keys, each containing "
-                "their respective service definition as JSON object.")
-QueryFilterServiceType = colander.SchemaNode(
-    colander.String(), name="type", missing=colander.drop,
-    example="api,thredds",
-    description="Comma-separated list of service-type for which to filter retrieved descriptions instead of all "
-                "available ones. Provided types matching is case insensitive. Unknown types are ignored.")
+class QueryEffectivePermissions(colander.MappingSchema):
+    effective = colander.SchemaNode(
+        colander.Boolean(), name="effective", default=False, missing=colander.drop,
+        description=(
+            "Obtain user's effective permissions resolved with corresponding service inheritance functionality. "
+            "(Note: Group inheritance is enforced regardless of 'inherited' query parameter values.)"
+        )
+    )
+
+
+class QueryInheritGroupsPermissions(colander.MappingSchema):
+    inherit = colander.SchemaNode(
+        colander.Boolean(), name="inherited", default=False, missing=colander.drop,
+        description=(
+            "Include the user's groups memberships inheritance to retrieve all possible permissions. "
+            "(Note: Duplicate, redundant and even conflicting permissions can be obtained when considering "
+            "multiple applied permissions individually applied from different group memberships or for the user "
+            "itself. See 'resolve' query parameter to reduce the set into a single highest priority permission)."
+        )
+    )
+
+
+class QueryResolvedUserGroupsPermissions(colander.MappingSchema):
+    resolve = colander.SchemaNode(
+        colander.Boolean(), name="resolve", default=False, missing=colander.drop,
+        description=(
+            "Combines corresponding direct user and groups inherited permissions into one, and locally resolves "
+            "for every resource the applicable permission modifiers considering group precedence and priorities. "
+            "(Note: Group permissions retrieval is enforced when using this option regardless of 'inherited' query "
+            "parameter since they are required to perform resolution)."
+        )
+    )
+
+
+class QueryFilterResources(colander.MappingSchema):
+    filtered = colander.SchemaNode(
+        colander.Boolean(), name="filtered", default=False, missing=colander.drop,
+        description=(
+            "Filter returned resources only where user has permissions on, either directly or inherited by groups "
+            "according to other query parameters. Otherwise (default), return all existing resources "
+            "with empty permission sets when user has no permission on them. Filtered view is enforced for "
+            "non-admin request user."
+        )
+    )
+
+
+class QueryParentResources(colander.MappingSchema):
+    parents = colander.SchemaNode(
+        colander.Boolean(), name="parents", default=False, missing=colander.drop,
+        description=(
+            "Obtain parent resources of the requested resource by ID instead of its children hierarchy. "
+            "The requested resource by ID will be the top-most object with nested 'parent' field "
+            "(instead of 'children' when 'parents=false') going all the way down to the root service of that resource. "
+            "Contrary to children hierarchy representation that can branch out over multiple children resources, "
+            "parent resources are guaranteed to have a single branch (each resource can only have one parent). "
+            "For this reason, it is possible to combine this parameter with 'flatten' and 'invert' to list resources "
+            "in a different order and/or format. These other queries are ignored when 'parents=false'."
+        )
+    )
+    flatten = colander.SchemaNode(
+        colander.Boolean(), name="flatten", default=False, missing=colander.drop,
+        description=(
+            "Applies only when 'parents=true'. Returns elements as a flattened list of JSON objects instead of default "
+            "response format as nested objects with resource ID keys and parent resource definitions under them. "
+            "Can be combined with 'invert' query to reverse the order of resource items in the returned list."
+        )
+    )
+    invert = colander.SchemaNode(
+        colander.Boolean(), name="invert", default=False, missing=colander.drop,
+        description=(
+            "Applies only when 'parents=true'. Returns elements in the inverse order, going from top-most service "
+            "down to (at most) the requested resource by ID listing all its intermediate children. This is different "
+            "than using 'parents=false' where the complete children hierarchy *under the requested resource ID* are "
+            "returned, with all possible branches. When this option is combined with 'parent', only the resources "
+            "on the specific branch directly linking the root service to the requested resource by ID are returned "
+            "instead of the full hierarchy, effectively only returning the resource "
+            "hierarchy *above the requested resource ID*. Note that for this reason, the requested resource will "
+            "always be the last/deepest item returned, even if it could have more children resources in reality. "
+            "Can be combined with 'flatten' query to obtain a list instead of nested objects."
+        )
+    )
+
+
+class QueryCascadeResourcesPermissions(colander.MappingSchema):
+    cascade = colander.SchemaNode(
+        colander.Boolean(), name="cascade", default=False, missing=colander.drop,
+        description=(
+            "Display all services that has at least one permission at any level in his hierarchy "
+            "(including all children resources). Otherwise (default), only returns services that have permissions "
+            "explicitly set on them, ignoring permissions set on children resources."
+        )
+    )
+
+
+class QueryFlattenServices(colander.MappingSchema):
+    flatten = colander.SchemaNode(
+        colander.Boolean(), name="flatten", default=False, missing=colander.drop,
+        description=(
+            "Return elements as a flattened list of JSON objects instead of default response format. "
+            "Default is a nested JSON of service-type keys with children service-name keys, each containing "
+            "their respective service definition as JSON object."
+        )
+    )
+
+
+class QueryFilterServiceType(colander.MappingSchema):
+    svc_type = colander.SchemaNode(
+        colander.String(), name="type", missing=colander.drop,
+        example="api,thredds",
+        description=(
+            "Comma-separated list of service-type for which to filter retrieved descriptions instead of all "
+            "available ones. Provided types matching is case insensitive. Unknown types are ignored."
+        )
+    )
 
 
 class PhoenixServicePushOption(colander.SchemaNode):
@@ -604,6 +681,7 @@ class ErrorCallBodySchema(ErrorFallbackBodySchema):
 
 class ErrorResponseBodySchema(BaseResponseBodySchema):
     def __init__(self, code, description, **kw):
+        # type: (int, str, Any) -> None
         super(ErrorResponseBodySchema, self).__init__(code, description, **kw)
         assert code >= 400  # nosec: B101
 
@@ -641,6 +719,7 @@ class ErrorResponseBodySchema(BaseResponseBodySchema):
 
 class InternalServerErrorResponseBodySchema(ErrorResponseBodySchema):
     def __init__(self, **kw):
+        # type: (Any) -> None
         kw["code"] = HTTPInternalServerError.code
         super(InternalServerErrorResponseBodySchema, self).__init__(**kw)
 
@@ -652,6 +731,7 @@ class BadRequestResponseSchema(BaseResponseSchemaAPI):
 
 class UnauthorizedResponseBodySchema(ErrorResponseBodySchema):
     def __init__(self, **kw):
+        # type: (Any) -> None
         kw["code"] = HTTPUnauthorized.code
         super(UnauthorizedResponseBodySchema, self).__init__(**kw)
 
@@ -1049,8 +1129,13 @@ class Resources_ResponseBodySchema(BaseResponseBodySchema):
     resources = ResourcesSchemaNode()
 
 
+class Resource_GET_RequestQuerySchema(QueryRequestSchemaAPI, QueryParentResources):
+    pass
+
+
 class Resource_GET_RequestSchema(BaseRequestSchemaAPI):
     path = Resource_RequestPathSchema()
+    querystring = Resource_GET_RequestQuerySchema()
 
 
 class Resource_MatchDictCheck_ForbiddenResponseSchema(BaseResponseSchemaAPI):
@@ -1069,6 +1154,8 @@ class Resource_MatchDictCheck_BadRequestResponseSchema(BaseResponseSchemaAPI):
 
 
 class Resource_GET_ResponseBodySchema(BaseResponseBodySchema):
+    # FIXME: OneOf required for alternatives (keys "parents" instead of "children", list representation)
+    #       When using list, 'resources' is used since many are under the same array field.
     resource = Resource_ParentResourceWithChildrenContainerBodySchema()
 
 
@@ -1078,7 +1165,7 @@ class Resource_GET_OkResponseSchema(BaseResponseSchemaAPI):
 
 
 class Resource_GET_InternalServerErrorResponseSchema(BaseResponseSchemaAPI):
-    description = "Failed building resource children json formatted tree."
+    description = "Failed building resource hierarchy into JSON formatted tree."
     body = InternalServerErrorResponseBodySchema(code=HTTPInternalServerError.code, description=description)
 
 
@@ -1343,8 +1430,8 @@ class ServiceTypesList(colander.SequenceSchema):
     )
 
 
-class ServiceListingQuerySchema(QueryRequestSchemaAPI):
-    flatten = QueryFlattenServices
+class ServiceListingQuerySchema(QueryRequestSchemaAPI, QueryFlattenServices):
+    pass
 
 
 class ServiceTypes_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -1404,9 +1491,8 @@ class Service_CheckConfig_UnprocessableEntityResponseSchema(BaseResponseSchemaAP
     body = ErrorResponseBodySchema(code=HTTPUnprocessableEntity.code, description=description)
 
 
-class ServicesQuerySchema(QueryRequestSchemaAPI):
-    flatten = QueryFlattenServices
-    svc_type = QueryFilterServiceType
+class ServicesQuerySchema(QueryRequestSchemaAPI, QueryFlattenServices, QueryFilterServiceType):
+    pass
 
 
 class Services_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2150,11 +2236,14 @@ class UserGroup_DELETE_NotFoundResponseSchema(BaseResponseSchemaAPI):
     body = ErrorResponseBodySchema(code=HTTPNotFound.code, description=description)
 
 
-class UserResources_GET_QuerySchema(QueryRequestSchemaAPI):
-    inherited = QueryInheritGroupsPermissions
-    resolve = QueryResolvedUserGroupsPermissions
-    filtered = QueryFilterResources
-    svc_type = QueryFilterServiceType
+class UserResources_GET_QuerySchema(
+    QueryRequestSchemaAPI,
+    QueryInheritGroupsPermissions,
+    QueryResolvedUserGroupsPermissions,
+    QueryFilterResources,
+    QueryFilterServiceType
+):
+    pass
 
 
 class UserResources_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2196,10 +2285,13 @@ class UserResourcePermissions_Check_ErrorResponseSchema(BaseResponseSchemaAPI):
     )
 
 
-class UserResourcePermissions_GET_QuerySchema(QueryRequestSchemaAPI):
-    inherited = QueryInheritGroupsPermissions
-    resolve = QueryResolvedUserGroupsPermissions
-    effective = QueryEffectivePermissions
+class UserResourcePermissions_GET_QuerySchema(
+    QueryRequestSchemaAPI,
+    QueryInheritGroupsPermissions,
+    QueryResolvedUserGroupsPermissions,
+    QueryEffectivePermissions
+):
+    pass
 
 
 class UserResourcePermissions_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2369,9 +2461,12 @@ class UserServiceResources_GET_OkResponseSchema(BaseResponseSchemaAPI):
     body = UserServiceResources_GET_ResponseBodySchema(code=HTTPOk.code, description=description)
 
 
-class UserServiceResources_GET_QuerySchema(QueryRequestSchemaAPI):
-    inherited = QueryInheritGroupsPermissions
-    resolve = QueryResolvedUserGroupsPermissions
+class UserServiceResources_GET_QuerySchema(
+    QueryRequestSchemaAPI,
+    QueryInheritGroupsPermissions,
+    QueryResolvedUserGroupsPermissions
+):
+    pass
 
 
 class UserServiceResources_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2413,11 +2508,14 @@ class UserServicePermissionName_DELETE_RequestSchema(BaseRequestSchemaAPI):
     body = colander.MappingSchema(default={})
 
 
-class UserServices_GET_QuerySchema(QueryRequestSchemaAPI):
-    cascade = QueryCascadeResourcesPermissions
-    inherit = QueryInheritGroupsPermissions
-    flatten = QueryFlattenServices
-    svc_type = QueryFilterServiceType
+class UserServices_GET_QuerySchema(
+    QueryRequestSchemaAPI,
+    QueryCascadeResourcesPermissions,
+    QueryInheritGroupsPermissions,
+    QueryFlattenServices,
+    QueryFilterServiceType
+):
+    pass
 
 
 class UserServices_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2434,9 +2532,12 @@ class UserServices_GET_OkResponseSchema(BaseResponseSchemaAPI):
     body = UserServices_GET_ResponseBodySchema(code=HTTPOk.code, description=description)
 
 
-class UserServicePermissions_GET_QuerySchema(QueryRequestSchemaAPI):
-    inherited = QueryInheritGroupsPermissions
-    resolve = QueryResolvedUserGroupsPermissions
+class UserServicePermissions_GET_QuerySchema(
+    QueryRequestSchemaAPI,
+    QueryInheritGroupsPermissions,
+    QueryResolvedUserGroupsPermissions
+):
+    pass
 
 
 class UserServicePermissions_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2643,8 +2744,8 @@ class GroupUsers_GET_ForbiddenResponseSchema(BaseResponseSchemaAPI):
     body = ErrorResponseBodySchema(code=HTTPForbidden.code, description=description)
 
 
-class GroupServices_GET_QuerySchema(QueryRequestSchemaAPI):
-    svc_type = QueryFilterServiceType
+class GroupServices_GET_QuerySchema(QueryRequestSchemaAPI, QueryFilterServiceType):
+    pass
 
 
 class GroupServices_GET_RequestSchema(BaseRequestSchemaAPI):
@@ -2816,8 +2917,8 @@ class GroupResourcePermissions_InternalServerErrorResponseSchema(BaseResponseSch
         code=HTTPInternalServerError.code, description=description)
 
 
-class GroupResources_GET_QuerySchema(QueryRequestSchemaAPI):
-    svc_type = QueryFilterServiceType
+class GroupResources_GET_QuerySchema(QueryRequestSchemaAPI, QueryFilterServiceType):
+    pass
 
 
 class GroupResources_GET_RequestSchema(BaseRequestSchemaAPI):